requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
idna==2.9
pkg-resources==0.0.0
psycopg2==2.8.5
pyTelegramBotAPI==3.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
